feature: Group tests
